mcp_server/requirements.txt
# IVD MCP Server Dependencies

# MCP protocol
mcp>=1.13.1

# HTTP server
starlette>=0.36.0
uvicorn>=0.27.0

# Embeddings
numpy>=1.26.0
openai>=1.12.0
tiktoken>=0.6.0

# YAML parsing
pyyaml>=6.0

# Logging
termcolor>=2.4.0
